pyproject.toml
[tool.poetry]
name = "mangashelf"
version = "0.1.0"
description = "SQLite organizer for manga collection with GUI"
authors = ["Jan Kaleta <kaletajan57@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
customtkinter = "^5.2.1"
pillow = "^10.1.0"
flake8 = "^6.1.0"
black = "^23.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
